requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose
opencv-python-headless
numpy
Pillow
tensorflow-cpu